requirements.txt
asgiref==3.8.1
attrs==25.3.0
bs4==0.0.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
dotenv==0.9.9
fonttools==4.58.4
h11==0.16.0
html5lib==1.1
idna==3.10
kiwisolver==1.4.8
lxml==6.0.0
Markdown==3.8.2
narwhals==1.45.0
numpy==2.3.1
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
plotly==6.2.0
pycparser==2.22
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
selenium==4.34.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
